zynqmp_pmufw: Fix violations for MISRA Rule 10_4
This patch fix following violation
misra_c_2012_rule_10_4_violation: Both operands of an operator in
which the usual arithmetic conversions are performed shall have the same
essential type category.

To solve this MISRA violation, explicitly type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_power.c
@@ -151,7 +151,7 @@ static PmNode* PmPowerDfsGetNext(void)
 {
 	PmNode* node = NULL;
 
-	while ((NULL != pmDfs.power) || (false == (u8)PmPowerStackIsEmpty())) {
+	while ((NULL != pmDfs.power) || (false == PmPowerStackIsEmpty())) {
 		if (NULL == pmDfs.power) {
 			PmPowerStackPop(&pmDfs.power, &pmDfs.it);
 		}
@@ -396,7 +396,7 @@ done:
  */
 static s32 PmPowerUpFpd(void)
 {
-	if (0 != (XPfw_Read32(PMU_GLOBAL_PWR_STATE) &
+	if (0U != (XPfw_Read32(PMU_GLOBAL_PWR_STATE) &
 				PMU_GLOBAL_PWR_STATE_FP_MASK)) {
 		PmInfo("Skipped FPD pwrup (FPD is on)\r\n");
 		return XST_SUCCESS;
